[update] move the mtimer implementation to aclint/mtimer.rs
